+++ covfcs/stk_nullcov.m
@@ -1,8 +1,10 @@
-% STK_COVMAT_NOISE [STK internal]
+% STK_NULLCOV ...
+%
+% FIXME: Document me!
 
 % Copyright Notice
 %
-%    Copyright (C) 2019 CentraleSupelec
+%    Copyright (C) 2021 CentraleSupelec
 %
 %    Author:  Julien Bect  <julien.bect@centralesupelec.fr>
 
@@ -11,7 +13,7 @@
 %    This file is part of
 %
 %            STK: a Small (Matlab/Octave) Toolbox for Kriging
-%               (https://github.com/stk-kriging/stk/)
+%               (http://sourceforge.net/projects/kriging)
 %
 %    STK is free software: you can redistribute it and/or modify it under
 %    the terms of the GNU General Public License as published by the Free
@@ -26,11 +28,25 @@
 %    You should  have received a copy  of the GNU  General Public License
 %    along with STK.  If not, see <http://www.gnu.org/licenses/>.
 
-function K = stk_covmat_noise (model, x1, x2, diff, pairwise)
+function K = stk_nullcov (param, x1, x2, diff, pairwise)  %#ok<INUSL>
 
-stk_error (['Classes derived from stk_model_ must implement ' ...
-    'stk_covmat_noise.'], 'IncompleteClassImplementation');
+% Number of evaluations points
+n1 = size (x1, 1);
+if (nargin > 2) && (~ isempty (x2))
+    n2 = size (x2, 1);
+else
+    n2 = n1;
+end
+
+% Default value for 'pairwise' (arg #5): false
+pairwise = (nargin > 4) && pairwise;
+assert ((n1 == n2) || (~ pairwise));
+
+% Return a matrix of zeros
+if pairwise
+    K = zeros (n1, 1);
+else
+    K = zeros (n1, n2);
+end
 
 end % function
-
-%#ok<*INUSD,*STOUT>