+++ lm/@stk_lm_/stk_lm_.m
@@ -1,8 +1,11 @@
-% STK_COVMAT_NOISE [STK internal]
+% STK_LM_ [internal]
+%
+% Base class for all lm objects.
+%
 
 % Copyright Notice
 %
-%    Copyright (C) 2019 CentraleSupelec
+%    Copyright (C) 2021 CentraleSupelec
 %
 %    Author:  Julien Bect  <julien.bect@centralesupelec.fr>
 
@@ -11,7 +14,7 @@
 %    This file is part of
 %
 %            STK: a Small (Matlab/Octave) Toolbox for Kriging
-%               (https://github.com/stk-kriging/stk/)
+%               (http://sourceforge.net/projects/kriging)
 %
 %    STK is free software: you can redistribute it and/or modify it under
 %    the terms of the GNU General Public License as published by the Free
@@ -26,11 +29,11 @@
 %    You should  have received a copy  of the GNU  General Public License
 %    along with STK.  If not, see <http://www.gnu.org/licenses/>.
 
-function K = stk_covmat_noise (model, x1, x2, diff, pairwise)
+function lm = stk_lm_ ()
 
-stk_error (['Classes derived from stk_model_ must implement ' ...
-    'stk_covmat_noise.'], 'IncompleteClassImplementation');
+lm = class (struct (), 'stk_lm_', stk_model_ ());
 
 end % function
 
-%#ok<*INUSD,*STOUT>
+
+%!test stk_test_class ('stk_lm_')