+++ lm/@stk_lm_noparam_/stk_get_optimizable_parameters.m
@@ -1,8 +1,17 @@
-% STK_COVMAT_NOISE [STK internal]
+% STK_GET_OPTIMIZABLE_PARAMETERS [overload STK function, internal]
+%
+% INTERNAL FUNCTION WARNING:
+%
+%    This function is currently considered as internal.  STK users that
+%    wish to experiment with parameter classes can already overload it,
+%    but should be aware that API-breaking changes are likely to happen
+%    in future releases.
+%
+% See also: stk_get_optimizable_parameters
 
 % Copyright Notice
 %
-%    Copyright (C) 2019 CentraleSupelec
+%    Copyright (C) 2021 CentraleSupelec
 %
 %    Author:  Julien Bect  <julien.bect@centralesupelec.fr>
 
@@ -11,7 +20,7 @@
 %    This file is part of
 %
 %            STK: a Small (Matlab/Octave) Toolbox for Kriging
-%               (https://github.com/stk-kriging/stk/)
+%               (http://sourceforge.net/projects/kriging)
 %
 %    STK is free software: you can redistribute it and/or modify it under
 %    the terms of the GNU General Public License as published by the Free
@@ -26,11 +35,8 @@
 %    You should  have received a copy  of the GNU  General Public License
 %    along with STK.  If not, see <http://www.gnu.org/licenses/>.
 
-function K = stk_covmat_noise (model, x1, x2, diff, pairwise)
+function value = stk_get_optimizable_parameters (lm)  %#ok<INUSD>
 
-stk_error (['Classes derived from stk_model_ must implement ' ...
-    'stk_covmat_noise.'], 'IncompleteClassImplementation');
+value = [];
 
 end % function
-
-%#ok<*INUSD,*STOUT>