+++ param/estim/stk_param_estim.m
@@ -36,7 +36,7 @@
 
 % Copyright Notice
 %
-%    Copyright (C) 2015-2018, 2020 CentraleSupelec
+%    Copyright (C) 2015-2018, 2020, 2021 CentraleSupelec
 %    Copyright (C) 2017 LNE
 %    Copyright (C) 2014 A. Ravisankar
 %    Copyright (C) 2011-2014 SUPELEC
@@ -66,179 +66,35 @@
 %    You should  have received a copy  of the GNU  General Public License
 %    along with STK.  If not, see <http://www.gnu.org/licenses/>.
 
-function [param_opt, lnv_opt, info] = stk_param_estim ...
-    (model, xi, zi, param0, lnv0, criterion)
+function [param_opt, lnv_opt, info] = stk_param_estim (varargin)
 
-stk_assert_model_struct (model);
-% FIXME: Implement stk_param_estim for noise priors too
+% TODO (?): Remove this function entirely, use stk_param_estim_ instead,
+%           and jump to major release 3.x
 
-% Empty is the same as 'not provided'
-if nargin < 6
-    criterion = [];
-    if nargin < 5
-        lnv0 = [];
-        if nargin < 4
-            param0 = [];
-        end
-    end
+switch nargout
+    
+    case {0, 1}
+        model_opt = stk_param_estim_ (varargin{:});
+        param_opt = model_opt.param;
+        
+    case 2
+        model_opt = stk_param_estim_ (varargin{:});
+        param_opt = model_opt.param;
+        lnv_opt = model_opt.lognoisevariance;
+        
+    case 3
+        [model_opt, info] = stk_param_estim_ (varargin{:});
+        param_opt = model_opt.param;
+        lnv_opt = model_opt.lognoisevariance;
+        
+    otherwise
+        stk_error ('Too many input arguments.', 'TooManyInputArgs');
+        
 end
-
-% Size checking: xi, zi
-zi_data = double (zi);
-if ~ isequal (size (zi_data), [stk_get_sample_size(xi) 1])
-    errmsg = 'zi should be a column, with the same number of rows as xi.';
-    stk_error (errmsg, 'IncorrectSize');
-end
-
-% Warn about special case: constant response
-if (std (zi_data) == 0)
-    warning ('STK:stk_param_estim:ConstantResponse', ['Constant-response ' ...
-        'data: the output of stk_param_estim is likely to be unreliable.']);
-end
-
-% Make sure that lognoisevariance is -inf for noiseless models
-if ~ stk_isnoisy (model)
-    model.lognoisevariance = -inf;
-end
-
-% Default criterion: restricted likelihood (ReML method)
-if isempty (criterion)
-    criterion = @stk_param_relik;
-end
-
-% Make sure that we have a starting point in (param0, lnv0)
-[param0, lnv0, do_estim_lnv] = provide_starting_point (model, xi, zi, param0, lnv0);
-
-% Set the starting point
-model.param = param0;
-model.lognoisevariance = lnv0;
-
-% Get selectors
-select = stk_param_getblockselectors (model);
-covparam_select = select{1};
-noiseparam_select = select{2} & do_estim_lnv;
-
-% NOTE ABOUT SELECTOrS / NaNs
-% The situation is currently a little odd, with a focus on estimating
-% covariance hyperparameters inherited from STK's history.  More precisely:
-%  * the 'linear model' part is not allowed to have (nonlinear) hyperparameters;
-%  * the hyperparameters of the covariance function are automatically estimated
-%    (all of them), regardless of whether there are NaNs or not in model.param;
-%  * the hyperparameters of the noise variance function are estimated (all of
-%    them) if there are NaNs in the vector OR if lnv0 is provided.  Note that
-%    only the case where lnv is a numerical scalar is currently documented.
-%    Also, note that all hyperparameters are estimated, even if only some of
-%    them are NaNs.
-% FIXME: Clarify this confusing situation...
-
-% Call optimization routine
-if nargout > 2
-    [model_opt, info] = stk_param_estim_optim ...
-        (model, xi, zi, criterion, covparam_select, noiseparam_select);
-else
-    model_opt = stk_param_estim_optim ...
-        (model, xi, zi, criterion, covparam_select, noiseparam_select);
-end
-
-param_opt = model_opt.param;
-lnv_opt = model_opt.lognoisevariance;
 
 end % function
 
 %#ok<*CTCH,*LERR,*SPWRN,*WNTAG>
-
-
-function [param0, lnv0, do_estim_lnv] = provide_starting_point ...
-    (model, xi, zi, param0, lnv0)
-
-% If param0 is not empty, it means that the user has provided
-% a starting point, in which case we must use it.
-if ~ isempty (param0)
-    
-    % Test if param0 contains NaNs of Infs
-    assert_is_finite (param0, 'param0');
-    % Cast to the appropriate type
-    if isnumeric (param0)
-        param0 = stk_set_optimizable_parameters (model.param, param0);
-    end
-    
-    % Now take care of lnv0.
-    % Same rule: if not empty, we have a user-provided starting point.
-    if isempty (lnv0)
-        % When lnv0 is not provided, noise variance estimation happens when lnv has NaNs.
-        do_estim_lnv = any (isnan (stk_get_optimizable_noise_parameters (model)));
-        if do_estim_lnv
-            % We have a user-provided starting point for param0 but not for lnv0.
-            model.param = param0;
-            lnv0 = stk_param_init_lnv (model, xi, zi);
-        else
-            lnv0 = model.lognoisevariance;
-        end
-    else
-        % When lnv0 is provided, noise variance *must* be estimated
-        do_estim_lnv = true;
-        % Test if lnv0 contains NaNs of Infs
-        assert_is_finite (lnv0, 'lnv0');
-        % Cast lnv0 to a variable of the appropriate type (numeric or object)
-        lnv0 = stk_set_optimizable_parameters (model.lognoisevariance, lnv0);
-    end
-    
-else  % Otherwise, try stk_param_init to get a starting point
-    
-    if isempty (lnv0)
-        % If needed, stk_param_init will also provide a starting point for lnv0
-        % (This is triggered by the presence of NaNs.)
-        do_estim_lnv = any (isnan (stk_get_optimizable_noise_parameters (model)));
-    else
-        % When lnv0 is provided, noise variance *must* be estimated
-        do_estim_lnv = true;
-        % Test if lnv0 contains NaNs of Infs
-        assert_is_finite (lnv0, 'lnv0');        
-        % Cast to the appropriate type
-        if isnumeric (lnv0)
-            lnv0 = stk_set_optimizable_parameters (model.lognoisevariance, param0);
-        end
-        % Install lnv0 in the model
-        model.lognoisevariance = lnv0;
-    end
-    
-    [param0, lnv0] = stk_param_init (model, xi, zi);
-    
-end % if
-
-end % function
-
-
-function assert_is_finite (param, param_name)
-
-% Check that there are no NaNs or Infs
-param_ = stk_get_optimizable_parameters (param);
-if ~ all (isfinite (param_))
-    stk_error (sprintf (['Incorrect value for input argument %s.  The components ' ...
-        'of the starting point must be neither infinite nor NaN.'], param_name),   ...
-        'InvalidArgument');
-end
-
-end % function
-
-
-function select = stk_param_getblockselectors (model)
-
-stk_assert_model_struct (model);
-
-select = cell (2, 1);
-
-% Covariance parameters
-covparam = stk_get_optimizable_parameters (model.param);
-covparam_size = length (covparam);
-select{1} = true (covparam_size, 1);
-
-% Noise parameters
-noiseparam = stk_get_optimizable_noise_parameters (model);
-noiseparam_size = length (noiseparam);
-select{2} = true (noiseparam_size, 1);
-
-end % function
 
 
 %!shared f, xi, zi, NI, param0, param1, model